refactor/ffi: safe_core and safe_app build and test working fine
